# Use string concatenation instead of list append