Performance is slightly better on arm64

#### before
```
     50000 calls,    5000000 bytes,     2112 ns/call
    100000 calls,   10000000 bytes,     2117 ns/call
```

#### after
```
     50000 calls,    5000000 bytes,     2067 ns/call
    100000 calls,   10000000 bytes,     2081 ns/call
```